+++ CHANGELOG.md
@@ -9,6 +9,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `product-summary-reference` block 
 
+## [2.77.1] - 2021-10-08
+
+### Added
+- Version param to images to permit caching it on CDN.
+
+## [2.77.0] - 2021-09-24
+
+### Added
+- I18n Bg and Bs.
+
+### Fixed
+- I18n En, Fr, It, Ko and Ro.
+
+## [2.76.0] - 2021-09-17
+### Added
+- The ability to dynamically choose the SKU that will be initially selected
+
+## [2.75.0] - 2021-08-23
+
+### Added
+- `asyncOnly1P` to `priceBehavior`.
+
+## [2.74.1] - 2021-08-09
+
+### Changed
+- Update default seller after the simulation call.
+
 ## [2.74.0] - 2021-06-28
 ### Changed
 - `listName` is sent via runtime params, not query string. This fixes the issue with other conflicting query strings.